+++ modules/hyp3_process/src/hyp3_process/daemon/__init__.py
@@ -1,1 +1,1 @@
-from .hyp3_daemon import HyP3Daemon, log, HyP3DaemonConfig
+from .daemon import HyP3Daemon, log, HyP3DaemonConfig
